[WIP] src:types: move netplan_net_definition into a separate file
This is to allow ABI compatibilty checks of NetplanNetDefinition using the
'abidiff' tool.
